Update pom.xml
The version of spring-boot-maven-plugin in your <plugins> section does not match the version of spring-boot-starter-parent in your <parent> section. They should be the same. You should change the version of spring-boot-maven-plugin to match the version of spring-boot-starter-parent, which is 3.1.5.

The spring-boot-maven-plugin is a plugin, not a dependency. It should be defined inside the <plugins> section of your pom.xml file, not in the <dependencies> section. You should remove the following lines from your <dependencies> section: